+++ packages/client/src/tag.rs
@@ -68,14 +68,14 @@ impl Tag {
 		self.string.as_str()
 	}
 
+	#[must_use]
 	pub fn name(&self) -> &str {
 		if let Some(Component::Version(_)) = self.components().last() {
 			let index = self
 				.string
 				.match_indices('/')
-				.last()
-				.map(|(index, _)| index)
-				.unwrap_or(self.string.len());
+				.next_back()
+				.map_or(self.string.len(), |(index, _)| index);
 			return &self.string[0..index];
 		}
 		self.as_str()
