+++ packages/server/src/checkin/input.rs
@@ -1,999 +1,491 @@
-use super::ignore;
-use crate::{Server, lockfile::ParsedLockfile};
-use futures::{TryStreamExt as _, stream::FuturesUnordered};
+use super::{
+	Directory, File, FileDependency, Node, State, Symlink, Variant, set_permissions_and_times,
+};
+use crate::Server;
 use std::{
-	collections::{BTreeMap, HashMap},
+	os::unix::fs::PermissionsExt as _,
 	path::{Path, PathBuf},
 	sync::Arc,
 };
 use tangram_client as tg;
 use tangram_either::Either;
-use tokio::sync::RwLock;
-
-#[derive(Debug)]
-pub struct Graph {
-	pub root: PathBuf,
-	pub nodes: Vec<Node>,
-}
-
-#[derive(Clone, Debug)]
-pub struct Node {
-	pub arg: tg::checkin::Arg,
-	pub edges: Vec<Edge>,
-	pub lockfile: Option<Arc<ParsedLockfile>>,
-	pub metadata: std::fs::Metadata,
-	pub parent: Option<usize>,
-}
-
-#[derive(Clone, Debug)]
-pub struct Edge {
-	pub kind: Option<tg::module::Kind>,
-	pub reference: tg::Reference,
-	pub subpath: Option<PathBuf>,
-	pub node: Option<usize>,
-	pub object: Option<tg::object::Id>,
-	pub path: Option<PathBuf>,
-	pub tag: Option<tg::Tag>,
-}
-
-struct State {
-	ignore_matcher: ignore::Matcher,
-	graph: Graph,
-	lockfiles: BTreeMap<PathBuf, Arc<ParsedLockfile>>,
-	visited: BTreeMap<PathBuf, usize>,
-}
 
 impl Server {
-	pub(super) async fn create_input_graph(
-		&self,
-		arg: tg::checkin::Arg,
-		progress: &crate::progress::Handle<tg::checkin::Output>,
-	) -> tg::Result<Graph> {
-		// Create the ignore matcher.
-		let ignore_matcher = Self::ignore_matcher_for_checkin().await?;
-
-		// Create the state.
-		let state = RwLock::new(State {
-			ignore_matcher,
-			lockfiles: BTreeMap::new(),
-			visited: BTreeMap::new(),
-			graph: Graph {
-				root: arg.path.clone(),
-				nodes: Vec::new(),
-			},
-		});
-
-		// Create the graph.
-		progress.start(
-			"files".into(),
-			"files".into(),
-			tg::progress::IndicatorFormat::Normal,
-			Some(0),
-			None,
-		);
-		Box::pin(self.create_input_graph_inner(None, arg.path.as_ref(), &arg, &state, progress))
-			.await?;
-		progress.finish("files");
-
-		// Get the graph.
-		let State { mut graph, .. } = state.into_inner();
-
-		// Find roots and subpaths.
-		progress.spinner("roots", "finding roots...");
-		graph.find_roots();
-		progress.finish("roots");
-
-		progress.spinner("subpaths", "fixing subpaths...");
-		graph.find_subpaths();
-		progress.finish("subpaths");
-
-		// Validate the graph.
-		progress.spinner("validate", "validating...");
-		graph.validate()?;
-		progress.finish("validate");
-
-		Ok(graph)
+	pub(super) fn checkin_collect_input(&self, state: &mut State, root: PathBuf) -> tg::Result<()> {
+		self.checkin_visit(state, root)?;
+		Self::checkin_find_roots(state);
+		Self::checkin_find_subpaths(state);
+		Ok(())
 	}
 
-	fn create_input_graph_inner<'a>(
-		&'a self,
-		referrer: Option<usize>,
-		path: &'a Path,
-		arg: &'a tg::checkin::Arg,
-		state: &'a RwLock<State>,
-		progress: &'a crate::progress::Handle<tg::checkin::Output>,
-	) -> impl Future<Output = tg::Result<usize>> + Send + 'a {
-		async move {
-			// Get the full path.
-			let path = if path.is_absolute() {
-				// If the absolute path is provided, use it.
-				path.to_owned()
-			} else {
-				let referrer = referrer.ok_or_else(|| tg::error!("expected a referrer"))?;
-				let referrer_path = state.read().await.graph.nodes[referrer].arg.path.clone();
-				referrer_path.join(path)
-			};
-
-			// Canonicalize the path.
-			let absolute_path = crate::util::fs::canonicalize_parent(&path).await.map_err(
-				|source| tg::error!(!source, %path = path.display(), "failed to canonicalize the parent"),
-			)?;
-
-			// Get the file system metadata.
-			let metadata = tokio::fs::symlink_metadata(&absolute_path).await.map_err(|source| tg::error!(!source, %path = absolute_path.display(), "failed to get file system metadata"))?;
-
-			// Validate.
-			if !(metadata.is_dir() || metadata.is_file() || metadata.is_symlink()) {
-				return Err(tg::error!(%path = absolute_path.display(), "invalid file type"));
-			}
-
-			// If this is a root module file, ensure the parent is collected.
-			if metadata.is_file() && tg::package::is_root_module_path(&absolute_path) {
-				let parent = absolute_path.parent().unwrap().to_owned();
-				Box::pin(self.create_input_graph_inner(None, &parent, arg, state, progress))
-				.await
-				.map_err(
-					|source| tg::error!(!source, %path = absolute_path.display(), "failed to collect input of parent"),
-				)?;
-			}
-
-			// Get a write lock on the state to avoid a race condition.
-			let mut state_ = state.write().await;
-
-			// Check if this path has already been visited and return it.
-			if let Some(node) = state_.visited.get(&absolute_path) {
-				return Ok(*node);
-			}
-
-			// Look up the parent if it exists.
-			let parent = absolute_path
-				.parent()
-				.and_then(|parent| state_.visited.get(parent).copied());
-
-			// Try to get the lockfile.
-			let lockfile = self
-				.try_get_lockfile(&absolute_path, &metadata, &mut state_)
-				.await?;
-
-			// Create a new node.
-			let arg_ = tg::checkin::Arg {
-				path: absolute_path.clone(),
-				..arg.clone()
-			};
-			let node = Node {
-				arg: arg_,
-				edges: Vec::new(),
-				lockfile,
-				metadata: metadata.clone(),
-				parent,
-			};
-			state_.graph.nodes.push(node);
-			let node = state_.graph.nodes.len() - 1;
-			state_.visited.insert(absolute_path.clone(), node);
-
-			// Get the list of dangling directories.
-			let dangling = state_.get_dangling_directories(node);
-
-			// Drop the state.
-			drop(state_);
-
-			// Collect any dangling directories.
-			for dangling in dangling {
-				Box::pin(self.create_input_graph_inner(None, &dangling, arg, state, progress))
-				.await
-				.map_err(
-					|source| tg::error!(!source, %path = dangling.display(), "failed to collect dangling directory"),
-				)?;
-			}
-
-			// Get the edges.
-			let edges = self
-				.get_edges(node, &absolute_path, arg, state, metadata, progress)
-				.await
-				.map_err(
-					|source| tg::error!(!source, %path = absolute_path.display(), "failed to get edges"),
-				)?;
-
-			// Update the node.
-			state.write().await.graph.nodes[node].edges = edges;
-
-			// Return the created node.
-			progress.increment("files", 1);
-			Ok(node)
-		}
-	}
-
-	async fn try_get_lockfile(
-		&self,
-		path: &Path,
-		metadata: &std::fs::Metadata,
-		state: &mut State,
-	) -> tg::Result<Option<Arc<ParsedLockfile>>> {
-		// First check if the lockfile is set by this file.
-		let lockfile_path = 'a: {
-			if !metadata.is_file() {
-				break 'a None;
-			}
-			let _permit = self.file_descriptor_semaphore.acquire().await.unwrap();
-			let Ok(Some(_)) = xattr::get(path, tg::file::XATTR_LOCK_NAME) else {
-				break 'a None;
-			};
-			Some(path.to_owned())
-		};
-
-		// Otherwise look up in the path's ancestors.
-		let lockfile_path = 'a: {
-			if let Some(path) = lockfile_path {
-				break 'a Some(path);
-			}
-			for path in path.ancestors() {
-				let path = path.join(tg::package::LOCKFILE_FILE_NAME);
-				if matches!(tokio::fs::try_exists(&path).await, Ok(true)) {
-					break 'a Some(path);
-				}
-			}
-			None
-		};
-
-		// If none is found, break early.
-		let Some(lockfile_path) = lockfile_path else {
-			return Ok(None);
-		};
-
-		// Check if a lockfile has already been read.
-		if let Some(lockfile) = state.lockfiles.get(&lockfile_path).cloned() {
-			return Ok(Some(lockfile));
+	fn checkin_visit(&self, state: &mut State, path: PathBuf) -> tg::Result<Option<usize>> {
+		// Check if the path has been visited.
+		if let Some(index) = state.graph.paths.get(&path) {
+			return Ok(Some(*index));
 		}
 
-		// Try to parse the lockfile.
-		let Ok(Some(lockfile)) = self.try_parse_lockfile(&lockfile_path) else {
-			return Ok(None);
-		};
-		let lockfile = Arc::new(lockfile);
-
-		// Update state.
-		state
-			.lockfiles
-			.insert(lockfile_path.clone(), lockfile.clone());
-
-		// Return the created lockfile.
-		Ok(Some(lockfile))
-	}
-
-	async fn get_edges(
-		&self,
-		referrer: usize,
-		path: &Path,
-		arg: &tg::checkin::Arg,
-		state: &RwLock<State>,
-		metadata: std::fs::Metadata,
-		progress: &crate::progress::Handle<tg::checkin::Output>,
-	) -> tg::Result<Vec<Edge>> {
-		if metadata.is_dir() {
-			Box::pin(self.get_directory_edges(referrer, path, arg, state, progress)).await
-		} else if metadata.is_file() {
-			Box::pin(self.get_file_edges(referrer, path, arg, state, progress)).await
-		} else if metadata.is_symlink() {
-			Box::pin(self.get_symlink_edges(referrer, path, arg, state, progress)).await
-		} else {
-			Err(tg::error!("invalid file type"))
-		}
-	}
-
-	async fn get_directory_edges(
-		&self,
-		referrer: usize,
-		path: &Path,
-		arg: &tg::checkin::Arg,
-		state: &RwLock<State>,
-		progress: &crate::progress::Handle<tg::checkin::Output>,
-	) -> tg::Result<Vec<Edge>> {
-		// Get the directory entries.
-		let mut names = Vec::new();
-		let permit = self.file_descriptor_semaphore.acquire().await.unwrap();
-		let mut read_dir = tokio::fs::read_dir(&path).await.map_err(
-			|source| tg::error!(!source, %path = path.display(), "failed to get the directory entries"),
+		// Get the metadata.
+		let metadata = std::fs::symlink_metadata(&path).map_err(
+			|source| tg::error!(!source, %path = path.display(), "failed to get the metadata"),
 		)?;
-		loop {
-			let Some(entry) = read_dir.next_entry().await.map_err(
-				|source| tg::error!(!source, %path = path.display(), "failed to get directory entry"),
-			)?
-			else {
-				break;
-			};
-			let name = entry.file_name().into_string().map_err(
-				|_| tg::error!(%path = path.display(), "directory contains entries with non-utf8 children"),
-			)?;
-			if arg.ignore {
-				let file_type = entry
-					.file_type()
-					.await
-					.map_err(|source| tg::error!(!source, "failed to get the file type"))?;
-				let is_directory = file_type.is_dir();
-				let ignored = state
-					.write()
-					.await
-					.ignore_matcher
-					.matches(&path.join(&name), Some(is_directory))
-					.await
-					.map_err(|source| {
-						tg::error!(!source, "failed to check if the path should be ignored")
-					})?;
-				if ignored {
-					continue;
-				}
-			}
+
+		// Skip ignored files.
+		if state.ignorer.as_mut().is_some_and(|ignorer| {
+			ignorer
+				.matches(&path, Some(metadata.is_dir()))
+				.unwrap_or_default()
+		}) {
+			return Ok(None);
+		}
+
+		// Create the variant.
+		let variant = if metadata.is_dir() {
+			Variant::Directory(Directory {
+				entries: Vec::new(),
+			})
+		} else if metadata.is_file() {
+			Variant::File(File {
+				blob: None,
+				dependencies: Vec::new(),
+				executable: metadata.permissions().mode() & 0o111 != 0,
+			})
+		} else if metadata.is_symlink() {
+			let target = std::fs::read_link(&path)
+				.map_err(|source| tg::error!(!source, "failed to read the symlink"))?;
+			Variant::Symlink(Symlink { target })
+		} else {
+			return Err(tg::error!(?metadata, "invalid file type"));
+		};
+
+		// If this is a destructive checkin, update permissions/times.
+		if state.arg.destructive {
+			set_permissions_and_times(&path, &metadata)?;
+		}
+
+		// Get the node index.
+		let index = state.graph.nodes.len();
+
+		// Update the path.
+		state.graph.paths.insert(path.clone(), index);
+
+		// Create the node.
+		let node = Node {
+			variant,
+			metadata: Some(metadata),
+			object: None,
+			path: Some(Arc::new(path)),
+			root: None,
+			tag: None,
+		};
+		state.graph.nodes.push_back(node);
+
+		// Visit the edges.
+		match &state.graph.nodes[index].variant {
+			Variant::Directory(_) => self.checkin_visit_directory_edges(state, index)?,
+			Variant::File(_) => self.checkin_visit_file_edges(state, index)?,
+			Variant::Symlink(_) => Self::checkin_visit_symlink_edges(state, index)?,
+			Variant::Object => return Err(tg::error!("unreachable")),
+		}
+
+		Ok(Some(index))
+	}
+
+	fn checkin_visit_directory_edges(&self, state: &mut State, index: usize) -> tg::Result<()> {
+		// Read the entries.
+		let read_dir = std::fs::read_dir(state.graph.nodes[index].path())
+			.map_err(|source| tg::error!(!source, "failed to read the directory"))?;
+		let mut names = Vec::new();
+		for result in read_dir {
+			let entry = result
+				.map_err(|source| tg::error!(!source, "failed to get the directory entry"))?;
+			let name = entry
+				.file_name()
+				.to_str()
+				.ok_or_else(|| tg::error!("expected the entry name to be a string"))?
+				.to_owned();
 			names.push(name);
 		}
-		drop(read_dir);
-		drop(permit);
 
-		let mut vec = Vec::with_capacity(names.len());
+		// Visit the children.
 		for name in names {
-			let path = path.join(&name);
-
-			// Follow the edge.
-			let node = Box::pin(self.create_input_graph_inner(
-				Some(referrer),
-				name.as_ref(),
-				arg,
-				state,
-				progress,
-			))
-			.await
-			.map_err(
-				|source| tg::error!(!source, %path = path.display(), "failed to collect child input"),
-			)?;
-
-			// Create the edge.
-			let reference = tg::Reference::with_path(&name);
-			let edge = Edge {
-				kind: None,
-				reference,
-				subpath: None,
-				node: Some(node),
-				object: None,
-				path: None,
-				tag: None,
-			};
-			vec.push(edge);
-		}
-		Ok(vec)
-	}
-
-	async fn get_file_edges(
-		&self,
-		referrer: usize,
-		path: &Path,
-		arg: &tg::checkin::Arg,
-		state: &RwLock<State>,
-		progress: &crate::progress::Handle<tg::checkin::Output>,
-	) -> tg::Result<Vec<Edge>> {
-		// Get the lockfile if it exists.
-		let lockfile = state.read().await.graph.nodes[referrer].lockfile.clone();
-
-		// Get the set of unresolved references.
-		let imports = 'a: {
-			// If this is a module path, attempt to parse it.
-			if !tg::package::is_module_path(path) {
-				break 'a None;
-			}
-
-			// Read the file, returning an error if we couldn't.
-			let permit = self.file_descriptor_semaphore.acquire().await.unwrap();
-			let text = tokio::fs::read_to_string(path).await.map_err(
-				|source| tg::error!(!source, %path = path.display(), "failed to read file contents"),
-			)?;
-			drop(permit);
-
-			// Analyze the module, forwarding errors as diagnostics.
-			let analysis = match crate::compiler::Compiler::analyze_module(text) {
-				Ok(analysis) => analysis,
-				Err(error) => {
-					// If analyzing the module fails, report a diagnostic and continue.
-					let diagnostic = tg::Diagnostic {
-						location: None,
-						severity: tg::diagnostic::Severity::Error,
-						message: format!("failed to analyze {}: {error}", path.display()),
-					};
-					progress.diagnostic(diagnostic);
-					break 'a None;
-				},
-			};
-
-			// Emit any errors as diagnostics.
-			for error in analysis.errors {
-				let diagnostic = tg::Diagnostic {
-					location: None,
-					severity: tg::diagnostic::Severity::Error,
-					message: format!("{error}"),
-				};
-				progress.diagnostic(diagnostic);
-			}
-
-			Some(analysis.imports)
-		};
-
-		// Pre-emptively pull tags for the set of imports.
-		if let Some(imports) = &imports {
-			let imports = imports.clone();
-			let server = self.clone();
-			tokio::spawn(async move {
-				server
-					.pull_tags_from_imports(imports)
-					.await
-					.inspect_err(|error| tracing::error!(%error, "failed to pull tags"))
-					.ok();
-			});
-		}
-
-		// Try and get the references of this file in the lockfile, if it exists.
-		let resolved_references = 'a: {
-			// Break if no lockfile is present.
-			let Some(lockfile) = lockfile else {
-				break 'a None;
-			};
-
-			// Break if we can't find the dependencies of this file.
-			let Ok(references) = lockfile.get_file_dependencies(path) else {
-				break 'a None;
-			};
-
-			// Collect the references from the lockfile.
-			let references = references
-				.into_iter()
-				.filter_map(|(reference, referent)| {
-					let referent = tg::Referent {
-						item: referent.item?,
-						path: referent.path,
-						subpath: referent.subpath,
-						tag: referent.tag,
-					};
-					Some((reference, referent))
-				})
-				.collect::<BTreeMap<_, _>>();
-			break 'a Some(references);
-		};
-
-		// Try and reconcile the resolved/unresolved references.
-		match (imports, resolved_references) {
-			// If both resolved references and imports are present, reconcile the two.
-			(Some(imports), Some(resolved_references)) => {
-				if arg.locked {
-					let matches = resolved_references.len() == imports.len()
-						&& imports
-							.iter()
-							.all(|import| resolved_references.contains_key(&import.reference));
-					if !matches {
-						return Err(tg::error!("lockfile is out of date"));
-					}
-				}
-
-				let mut edges = Vec::with_capacity(imports.len());
-				for import in imports {
-					let edge = if let Some(referent) = resolved_references.get(&import.reference) {
-						self.get_edge_from_path_or_id(
-							import.reference.clone(),
-							referent.clone(),
-							referrer,
-							path,
-							arg,
-							state,
-							progress,
-						)
-						.await?
-					} else {
-						self.get_edge_from_import(import, referrer, path, arg, state, progress)
-							.await?
-					};
-					edges.push(edge);
-				}
-
-				// Return the merged references
-				Ok(edges)
-			},
-
-			// If only unresolved references exist, return them.
-			(Some(imports), None) => {
-				let mut edges = Vec::with_capacity(imports.len());
-				for import in imports {
-					let edge = self
-						.get_edge_from_import(import, referrer, path, arg, state, progress)
-						.await?;
-					edges.push(edge);
-				}
-				Ok(edges)
-			},
-
-			// If only resolved references exist, return them.
-			(None, Some(resolved_references)) => {
-				let mut edges = Vec::with_capacity(resolved_references.len());
-				for (reference, referent) in resolved_references {
-					let edge = self
-						.get_edge_from_path_or_id(
-							reference, referent, referrer, path, arg, state, progress,
-						)
-						.await?;
-					edges.push(edge);
-				}
-				Ok(edges)
-			},
-
-			// Otherwise return no references.
-			(None, None) => Ok(Vec::new()),
-		}
-	}
-
-	async fn get_edge_from_import(
-		&self,
-		import: tg::module::Import,
-		referrer: usize,
-		path: &Path,
-		arg: &tg::checkin::Arg,
-		state: &RwLock<State>,
-		progress: &crate::progress::Handle<tg::checkin::Output>,
-	) -> tg::Result<Edge> {
-		// Follow path dependencies.
-		let import_path = import
-			.reference
-			.item()
-			.try_unwrap_path_ref()
-			.ok()
-			.or_else(|| import.reference.options()?.path.as_ref());
-		if let Some(import_path) = import_path {
-			// Create the reference.
-			let reference = import.reference.clone();
-
-			// Get the absolute path of the referent.
-			let absolute_path = crate::util::fs::canonicalize_parent(path.parent().unwrap().join(import_path)).await.map_err(|source| tg::error!(!source, %path = path.display(), "failed to canonicalize the path's parent"))?;
-
-			// Get the referent.
-			let child = self
-				.create_input_graph_inner(Some(referrer), &absolute_path, arg, state, progress)
-				.await
-				.map_err(|source| tg::error!(!source, "failed to collect child input"))?;
-
-			// If the child is a package and the import kind is not directory, get the subpath.
-			let (child_path, is_directory) = {
-				let state = state.read().await;
-				(
-					state.graph.nodes[child].arg.path.clone(),
-					state.graph.nodes[child].metadata.is_dir(),
-				)
-			};
-
-			// Get the subpath if this is a package import.
-			let (node, subpath) = if matches!(&import.kind, Some(tg::module::Kind::Directory))
-				|| !is_directory
-			{
-				(child, None)
-			} else {
-				let subpath = tg::package::try_get_root_module_file_name(self,Either::Right(&child_path)).await.map_err(|source| tg::error!(!source, %path = child_path.display(), "failed to get root module file name"))?
-						.map(PathBuf::from);
-				(child, subpath)
-			};
-
-			// Create the edge.
-			let package_path = state.read().await.graph.nodes[node].arg.path.clone();
-			let path = crate::util::path::diff(&arg.path, &package_path)?;
-			let edge = Edge {
-				kind: import.kind,
-				reference,
-				node: Some(node),
-				object: None,
-				path: Some(path),
-				subpath,
-				tag: None,
-			};
-			return Ok(edge);
-		}
-
-		Ok(Edge {
-			kind: import.kind,
-			reference: import.reference.clone(),
-			subpath: None,
-			node: None,
-			object: None,
-			path: None,
-			tag: None,
-		})
-	}
-
-	#[allow(clippy::too_many_arguments)]
-	async fn get_edge_from_path_or_id(
-		&self,
-		reference: tg::Reference,
-		referent: tg::Referent<Either<PathBuf, tg::object::Id>>,
-		referrer: usize,
-		_path: &Path,
-		arg: &tg::checkin::Arg,
-		state: &RwLock<State>,
-		progress: &crate::progress::Handle<tg::checkin::Output>,
-	) -> tg::Result<Edge> {
-		let node = if let Either::Left(path) = &referent.item {
-			let node =
-				Box::pin(self.create_input_graph_inner(Some(referrer), path, arg, state, progress))
-					.await?;
-			Some(node)
-		} else {
-			None
-		};
-		let object = if let Either::Right(object) = &referent.item {
-			Some(object.clone())
-		} else {
-			None
-		};
-		let edge = Edge {
-			kind: None,
-			reference,
-			subpath: referent.subpath,
-			node,
-			object,
-			path: referent.path,
-			tag: referent.tag,
-		};
-		Ok(edge)
-	}
-
-	async fn get_symlink_edges(
-		&self,
-		referrer: usize,
-		path: &Path,
-		arg: &tg::checkin::Arg,
-		state: &RwLock<State>,
-		_progress: &crate::progress::Handle<tg::checkin::Output>,
-	) -> tg::Result<Vec<Edge>> {
-		// Check if this node's edges have already been created.
-		let existing = state.read().await.graph.nodes[referrer].edges.clone();
-		if !existing.is_empty() {
-			return Ok(existing);
-		}
-
-		// Read the symlink.
-		let target = tokio::fs::read_link(path).await.map_err(
-			|source| tg::error!(!source, %path = path.display(), "failed to read symlink"),
-		)?;
-
-		// Error if this is an absolute path or empty.
-		if target.is_absolute() {
-			return Err(
-				tg::error!(%path = path.display(), %target = target.display(), "cannot check in an absolute path symlink"),
-			);
-		} else if target.as_os_str().is_empty() {
-			return Err(
-				tg::error!(%path = path.display(),  "cannot check in a path with an empty target"),
-			);
-		}
-
-		// Get the full target by joining with the parent.
-		let target_absolute_path = crate::util::fs::canonicalize_parent(
-			path.parent().unwrap().join(&target),
-		)
-		.await
-		.map_err(
-			|source| tg::error!(!source, %path = path.display(), "failed to canonicalize the path's parent"),
-		)?;
-
-		// Check if this is a checkin of an artifact.
-		let artifacts_path = self.get_artifacts_path_for_path(path).await?;
-		let path = crate::util::path::diff(&arg.path, &target_absolute_path)?;
-		let edge = if let Ok(subpath) = target_absolute_path.strip_prefix(artifacts_path) {
-			let mut components = subpath.components();
-			let object = components
-				.next()
-				.ok_or_else(
-					|| tg::error!(%subpath = subpath.display(), "expected a subpath component"),
-				)?
-				.as_os_str()
-				.to_str()
-				.ok_or_else(|| tg::error!(%subpath = subpath.display(), "invalid subpath"))?
-				.parse()
-				.map_err(|_| tg::error!(%subpath = subpath.display(), "expected an object id"))?;
-			let subpath: PathBuf = components.collect();
-			Edge {
-				kind: None,
-				reference: tg::Reference::with_path(target),
-				subpath: Some(subpath),
-				node: None,
-				object: Some(object),
-				path: Some(path),
-				tag: None,
-			}
-		} else {
-			Edge {
-				kind: None,
-				reference: tg::Reference::with_path(target),
-				subpath: None,
-				node: None,
-				object: None,
-				path: Some(path),
-				tag: None,
-			}
-		};
-		Ok(vec![edge])
-	}
-
-	async fn get_artifacts_path_for_path(&self, path: &Path) -> tg::Result<PathBuf> {
-		let mut artifacts_path = None;
-		for ancestor in path.ancestors().skip(1) {
-			let path = ancestor.join(".tangram/artifacts");
-			if matches!(tokio::fs::try_exists(&path).await, Ok(true)) {
-				artifacts_path.replace(path);
-				break;
-			}
-		}
-		Ok(artifacts_path.unwrap_or(self.artifacts_path()))
-	}
-}
-
-impl Server {
-	async fn pull_tags_from_imports(
-		&self,
-		imports: impl IntoIterator<Item = tg::module::Import>,
-	) -> tg::Result<()> {
-		imports
-			.into_iter()
-			.filter_map(|import| {
-				let server = self.clone();
-				let remote = import
-					.reference
-					.options()
-					.and_then(|options| options.remote.clone());
-				let pattern = import.reference.item().try_unwrap_tag_ref().ok()?.clone();
-				let future = async move { server.pull_tag(pattern, remote).await };
-				Some(future)
-			})
-			.collect::<FuturesUnordered<_>>()
-			.try_collect()
-			.await
-	}
-}
-
-impl Graph {
-	// Find the roots of every node.
-	fn find_roots(&mut self) {
-		// Build a HashMap of paths to node indices.
-		let mut path_to_node: HashMap<PathBuf, usize> = HashMap::new();
-		for (idx, node) in self.nodes.iter().enumerate() {
-			path_to_node.insert(node.arg.path.clone(), idx);
-		}
-		'outer: for node in 0..self.nodes.len() {
-			// Find an ancestor using the path HashMap
-			let ancestor = {
-				let node_path = &self.nodes[node].arg.path;
-				let mut current_path = node_path.as_path();
-				let mut result = None;
-
-				// Walk up the path hierarchy
-				while let Some(parent_path) = current_path.parent() {
-					if let Some(&idx) = path_to_node.get(parent_path) {
-						// Skip non-directories
-						if self.nodes[idx].metadata.is_dir() && idx != node {
-							result = Some(idx);
-							break;
-						}
-					}
-					current_path = parent_path;
-				}
-				result
-			};
-
-			// If there are no ancestors of this node, strip its parent.
-			let Some(ancestor) = ancestor else {
-				self.nodes[node].parent.take();
-				continue 'outer;
-			};
-
-			// Otherwise, walk the ancestor's children to find the parent of the node.
-			let mut stack = vec![ancestor];
-			while let Some(parent) = stack.pop() {
-				for child in self.nodes[parent].edges.iter().filter_map(|edge| edge.node) {
-					if child == node {
-						self.nodes[node].parent.replace(parent);
-						continue 'outer;
-					}
-					if self.nodes[child].metadata.is_dir() {
-						stack.push(child);
-					}
-				}
-			}
-		}
-	}
-
-	// Convert dependencies into root + subpath.
-	fn find_subpaths(&mut self) {
-		for node in 0..self.nodes.len() {
-			// Only consider module imports.
-			if !(self.nodes[node].metadata.is_file()
-				&& tg::package::is_module_path(&self.nodes[node].arg.path))
-			{
-				continue;
-			}
-
-			// Remap edges.
-			for edge_index in 0..self.nodes[node].edges.len() {
-				// Skip edges that don't point to inputs.
-				let Some(referent) = self.nodes[node].edges[edge_index].node else {
-					continue;
-				};
-
-				// Skip edgs that have their subpath already set.
-				if self.nodes[node].edges[edge_index].subpath.is_some() {
-					continue;
-				}
-
-				// Get the root and subpath of the referent.
-				let (package, subpath) = root_node_with_subpath(self, referent);
-
-				// Get the path of the root.
-				let path = crate::util::path::diff(&self.root, &self.nodes[package].arg.path).ok();
-
-				// Create a new edge.
-				let edge = Edge {
-					kind: self.nodes[node].edges[edge_index].kind,
-					node: Some(package),
-					reference: self.nodes[node].edges[edge_index].reference.clone(),
-					subpath,
-					object: None,
-					path,
-					tag: None,
-				};
-
-				// Update the edge.
-				self.nodes[node].edges[edge_index] = edge;
-			}
-		}
-	}
-
-	// Validate the input graph:
-	fn validate(&self) -> tg::Result<()> {
-		// Get the paths of all the nodes via DFS walk of the graph.
-		let mut paths = BTreeMap::new();
-		let mut stack = vec![0];
-		while let Some(node) = stack.pop() {
-			let path = self.nodes[node].arg.path.clone();
-			if paths.contains_key(&path) {
-				continue;
-			}
-			paths.insert(path, node);
-			for edge in &self.nodes[node].edges {
-				if let Some(child) = edge.node {
-					stack.push(child);
-				}
-			}
-		}
-
-		// Validate that all parents are reachable.
-		for (path, node) in &paths {
-			let Some(parent) = self.nodes[*node].parent else {
+			let path = state.graph.nodes[index].path().join(&name);
+			let Some(child_index) = self.checkin_visit(state, path)? else {
 				continue;
 			};
-			let parent = self.nodes[parent].arg.path.clone();
-			let _parent = paths.get(&parent).ok_or_else(
-				|| tg::error!(%path = path.display(), %parent = parent.display(), "missing parent node"),
-			)?;
+			state.graph.nodes[index]
+				.variant
+				.unwrap_directory_mut()
+				.entries
+				.push((name, child_index));
 		}
 
-		// Validate that no modules imported by other modules are outside of the same root.
-		for (index, node) in self.nodes.iter().enumerate() {
-			if !node.metadata.is_file() {
-				continue;
-			}
+		Ok(())
+	}
 
-			// Get the root of this node.
-			let root = get_root_node(self, index);
-			for edge in &node.edges {
-				let Some(child) = edge.node else {
-					continue;
-				};
-				if !self.nodes[child].metadata.is_file()
-					|| !tg::package::is_module_path(&self.nodes[child].arg.path)
+	fn checkin_visit_file_edges(&self, state: &mut State, index: usize) -> tg::Result<()> {
+		// Get the list of all dependencies.
+		let path = state.graph.nodes[index].path().to_owned();
+		let mut dependencies = self.get_file_dependencies(state, &path)?;
+
+		// Visit path dependencies.
+		for dependency in &mut dependencies {
+			match dependency {
+				FileDependency::Import { import, node, .. }
+					if import.reference.path().is_some() =>
 				{
-					continue;
+					let path = path
+						.parent()
+						.unwrap()
+						.join(import.reference.path().unwrap());
+					let path = crate::util::fs::canonicalize_parent_sync(&path).map_err(
+						|source| tg::error!(!source, %path = path.display(), "failed to canonicalize path"),
+					)?;
+					let Some(index) = self.checkin_visit(state, path)? else {
+						continue;
+					};
+					*node = Some(index);
+				},
+				_ => (),
+			}
+		}
+
+		// Update the graph.
+		state.graph.nodes[index]
+			.variant
+			.unwrap_file_mut()
+			.dependencies = dependencies;
+
+		Ok(())
+	}
+
+	fn get_file_dependencies(
+		&self,
+		state: &mut State,
+		path: &Path,
+	) -> tg::Result<Vec<FileDependency>> {
+		// Check if this file has dependencies set in the xattr.
+		if let Ok(Some(contents)) = xattr::get(path, tg::file::XATTR_LOCK_NAME) {
+			let lockfile = serde_json::from_slice::<tg::Lockfile>(&contents)
+				.map_err(|source| tg::error!(!source, "failed to deserialize lockfile"))?;
+			if lockfile.nodes.len() != 1 {
+				return Err(tg::error!(%path = path.display(), "expected single node in lockfile"));
+			}
+			let Some(tg::lockfile::Node::File(file_node)) = lockfile.nodes.first() else {
+				return Err(tg::error!(%path = path.display(), "expected a file node"));
+			};
+			return file_node.dependencies.iter().try_fold(
+				Vec::new(),
+				|mut acc, (reference, referent)| match &referent.item {
+					Either::Left(_) => Err(tg::error!("found a graph node")),
+					Either::Right(object) => {
+						let referent = tg::Referent {
+							item: Some(Either::Left(object.clone())),
+							path: referent.path.clone(),
+							subpath: referent.subpath.clone(),
+							tag: referent.tag.clone(),
+						};
+						let dependency = FileDependency::Referent {
+							reference: reference.clone(),
+							referent,
+						};
+						acc.push(dependency);
+						Ok(acc)
+					},
+				},
+			);
+		}
+
+		// If this is not a module, it has no dependencies.
+		if !tg::package::is_module_path(path) {
+			return Ok(Vec::new());
+		}
+
+		// Parse imports.
+		let contents = std::fs::read(path).map_err(
+			|source| tg::error!(!source, %path = path.display(), "failed to read module file"),
+		)?;
+		let text = String::from_utf8(contents)
+			.map_err(|source| tg::error!(!source, %path = path.display(), "invalid utf8"))?;
+		let analysis = crate::Compiler::analyze_module(text)?;
+		for error in analysis.errors {
+			// Dump diagnostics.
+			let diagnostic = tg::Diagnostic {
+				location: None,
+				message: error.to_string(),
+				severity: tg::diagnostic::Severity::Error,
+			};
+			state.progress.diagnostic(diagnostic);
+		}
+
+		// Get the locked dependencies.
+		let locked_dependencies = state
+			.lockfile
+			.as_ref()
+			.and_then(|lockfile| lockfile.get_file_dependencies(path).ok())
+			.unwrap_or_default();
+
+		// Make sure that the lockfile dependencies matches the import set.
+		if state.arg.locked
+			&& (locked_dependencies.len() != analysis.imports.len()
+				|| !locked_dependencies.iter().all(|(reference, _)| {
+					analysis
+						.imports
+						.iter()
+						.any(|import| &import.reference == reference)
+				})) {
+			return Err(tg::error!(
+				"the lockfile needs to be updated but --locked was passed"
+			));
+		}
+
+		let dependencies =
+			analysis
+				.imports
+				.into_iter()
+				.try_fold(Vec::new(), |mut acc, import| {
+					// Use the locked dependency if the reference matches.
+					if let Some((reference, referent)) =
+						locked_dependencies
+							.iter()
+							.find_map(|(reference, referent)| {
+								if &import.reference != reference {
+									return None;
+								}
+								let item = referent.item.as_ref()?.as_ref().right()?.clone();
+								let referent = tg::Referent {
+									item: Some(Either::Left(item)),
+									path: referent.path.clone(),
+									subpath: referent.subpath.clone(),
+									tag: referent.tag.clone(),
+								};
+								Some((reference.clone(), referent))
+							}) {
+						acc.push(FileDependency::Referent {
+							reference,
+							referent,
+						});
+						return Ok(acc);
+					}
+
+					// Return an error if this dependency can't be resolved.
+					if state.locked
+						&& import.reference.path().is_none()
+						&& import.reference.item().try_unwrap_tag_ref().is_ok()
+					{
+						return Err(
+							tg::error!(%import = &import.reference, "unresolved import when --locked"),
+						);
+					}
+
+					// Pull tags.
+					if let Ok(pattern) = import.reference.item().try_unwrap_tag_ref() {
+						tokio::spawn({
+							let server = self.clone();
+							let pattern = pattern.clone();
+							let remote = import
+								.reference
+								.options()
+								.and_then(|options| options.remote.clone());
+							async move {
+								server.pull_tag(pattern.clone(), remote.clone()).await.ok();
+							}
+						});
+					}
+
+					// Get the subpath of this import if it exists. Since this may change later, it's kept separate to preserve the original import reference/options.
+					let subpath = import
+						.reference
+						.options()
+						.and_then(|opt| opt.subpath.clone());
+
+					// Add the import.
+					acc.push(FileDependency::Import {
+						import,
+						node: None,
+						path: None,
+						subpath,
+					});
+
+					Ok(acc)
+				})?;
+
+		Ok(dependencies)
+	}
+
+	#[allow(clippy::unnecessary_wraps)]
+	fn checkin_visit_symlink_edges(_state: &mut State, _index: usize) -> tg::Result<()> {
+		// TODO
+		Ok(())
+	}
+
+	fn checkin_find_roots(state: &mut State) {
+		let mut visited = vec![false; state.graph.nodes.len()];
+		let mut stack = vec![(0, None::<usize>)];
+		'outer: while let Some((index, hint)) = stack.pop() {
+			// Skip nodes that have already been visited.
+			if visited[index] {
+				continue;
+			}
+			visited[index] = true;
+
+			// Walk up the path hierarchy.
+			let path = state.graph.nodes[index].path();
+
+			// Check if the hint matches.
+			if let (Some(hint), false) = (hint, state.graph.nodes[index].is_package()) {
+				let hint_path = state.graph.nodes[hint].path();
+				if path.strip_prefix(hint_path).is_ok() {
+					// Mark this node's root.
+					state.graph.nodes[index].root.replace(hint);
+					state.graph.roots.entry(hint).or_default().push(index);
+
+					// Recurse on children using the same hint.
+					let children = state.graph.nodes[index]
+						.edges()
+						.into_iter()
+						.map(|child| (child, Some(hint)));
+					stack.extend(children);
+					continue 'outer;
 				}
-				let root_of_child = get_root_node(self, child);
-				if root_of_child != root {
-					return Err(tg::error!(
-						%path = node.arg.path.display(),
-						%reference = edge.reference,
-						"cannot import external modules"
-					));
+			}
+
+			// Search the ancestors of this node to find its root, unless it is a package.
+			let mut root = None;
+			if !state.graph.nodes[index].is_package() {
+				for ancestor in path.ancestors().skip(1) {
+					let Some(node) = state.graph.paths.get(ancestor) else {
+						break;
+					};
+					root.replace(*node);
 				}
 			}
+
+			// Add to the list of roots if necessary.
+			if let Some(root) = root {
+				state.graph.roots.entry(root).or_default().push(index);
+				state.graph.nodes[index].root.replace(root);
+			} else {
+				state.graph.roots.entry(index).or_default();
+			}
+
+			// Recurse, using the root if it was discovered or this node if it is a directory.
+			let hint = Some(root.unwrap_or(index));
+			let children = state.graph.nodes[index]
+				.edges()
+				.into_iter()
+				.map(|child| (child, hint));
+
+			stack.extend(children);
 		}
-
-		Ok(())
-	}
-}
-
-fn root_node_with_subpath(graph: &Graph, node: usize) -> (usize, Option<PathBuf>) {
-	// Find the root node.
-	let root = get_root_node(graph, node);
-
-	// If this is a root node, return it.
-	if root == node {
-		return (node, None);
 	}
 
-	// Otherwise compute the subpath within the root.
-	let referent_path = graph.nodes[node].arg.path.clone();
-	let root_path = graph.nodes[root].arg.path.clone();
-	let subpath = referent_path.strip_prefix(&root_path).unwrap().to_owned();
+	fn checkin_find_subpaths(state: &mut State) {
+		for index in 0..state.graph.nodes.len() {
+			if !(state.graph.nodes[index]
+				.metadata
+				.as_ref()
+				.unwrap()
+				.is_file() && tg::package::is_module_path(state.graph.nodes[index].path()))
+			{
+				continue;
+			}
 
-	// Return the root node and subpath.
-	(root, Some(subpath))
-}
+			let updates = state.graph.nodes[index]
+				.variant
+				.unwrap_file_ref()
+				.dependencies
+				.iter()
+				.enumerate()
+				.filter_map(|(dep_index, dependency)| {
+					// Skip unresolved dependencies.
+					let FileDependency::Import {
+						import,
+						node: Some(node),
+						subpath: None,
+						..
+					} = dependency
+					else {
+						return None;
+					};
 
-impl State {
-	fn get_dangling_directories(&mut self, node: usize) -> Vec<PathBuf> {
-		// Get the root.
-		let root = get_root_node(&self.graph, node);
+					// Skip any imports that have subpaths.
+					if import
+						.reference
+						.options()
+						.is_some_and(|options| options.subpath.is_some())
+					{
+						return None;
+					}
 
-		// Get the paths of this node and the root.
-		let node_path = &self.graph.nodes[node].arg.path;
-		let root_path = &self.graph.nodes[root].arg.path;
+					// Get the package and subpath of the import.
+					let (package, subpath) = state.graph.nodes[*node].root.map_or_else(
+						|| {
+							// If the import kind is _not_ a module, make sure to ignore the result.
+							if !matches!(
+								import.kind,
+								None | Some(
+									tg::module::Kind::Dts
+										| tg::module::Kind::Js | tg::module::Kind::Ts
+								)
+							) {
+								return (*node, None);
+							}
 
-		// Check if this is a child of the root.
-		let Ok(diff) = root_path.strip_prefix(node_path) else {
-			return Vec::new();
-		};
+							// If this is a root directory and contains a tangram.ts, use it.
+							let subpath = state.graph.nodes[*node]
+								.variant
+								.try_unwrap_directory_ref()
+								.ok()
+								.and_then(|directory| {
+									directory.entries.iter().find_map(|(name, _)| {
+										tg::package::ROOT_MODULE_FILE_NAMES
+											.contains(&name.as_str())
+											.then_some(name.into())
+									})
+								});
+							(*node, subpath)
+						},
+						|root| {
+							// Otherwise look for this as a subpath of the module.
+							let subpath = state.graph.nodes[*node]
+								.path()
+								.strip_prefix(state.graph.nodes[root].path())
+								.unwrap()
+								.to_owned();
+							(root, Some(subpath))
+						},
+					);
 
-		// Collect any paths along the way that haven't been visited.
-		let mut dangling = Vec::new();
-		let mut path = root_path.to_owned();
-		for component in diff.components() {
-			path = path.join(component);
-			if !self.visited.contains_key(&path) {
-				dangling.push(path.clone());
+					// Use the relative path of the package
+					let path = state.graph.nodes[package]
+						.path
+						.as_deref()
+						.and_then(|module_path| {
+							let path =
+								crate::util::path::diff(&state.arg.path, module_path).ok()?;
+							(!path.as_os_str().is_empty()).then_some(path)
+						});
+
+					// Recreate the new dependency.
+					Some((
+						dep_index,
+						FileDependency::Import {
+							import: import.clone(),
+							node: Some(package),
+							path,
+							subpath,
+						},
+					))
+				})
+				.collect::<Vec<_>>(); // iterator ivnalidation.
+
+			// Replace the import with the changed one.
+			for (dep_index, dependency) in updates {
+				state.graph.nodes[index]
+					.variant
+					.unwrap_file_mut()
+					.dependencies[dep_index] = dependency;
 			}
 		}
-
-		dangling
-	}
-}
-
-fn get_root_node(graph: &Graph, mut node: usize) -> usize {
-	loop {
-		let Some(parent) = graph.nodes[node].parent else {
-			return node;
-		};
-		node = parent;
-	}
-}
-
-impl std::fmt::Display for Graph {
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		for (index, node) in self.nodes.iter().enumerate() {
-			let path = self.nodes[index].arg.path.clone();
-			let path = crate::util::path::diff(&self.root, &path).unwrap();
-			writeln!(f, "{index} {}", path.display())?;
-			for edge in &node.edges {
-				writeln!(f, "\t{edge}")?;
-			}
-		}
-		Ok(())
-	}
-}
-
-impl std::fmt::Display for Edge {
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		write!(f, "(reference: {}", self.reference)?;
-		if let Some(node) = &self.node {
-			write!(f, ", node: {node}")?;
-		}
-		if let Some(object) = &self.object {
-			write!(f, ", object: {object}")?;
-		}
-		if let Some(path) = &self.path {
-			write!(f, ", path: {}", path.display())?;
-		}
-		if let Some(subpath) = &self.subpath {
-			write!(f, ", subpath: {}", subpath.display())?;
-		}
-		if let Some(tag) = &self.tag {
-			write!(f, ", tag: {tag}")?;
-		}
-		write!(f, ")")?;
-		Ok(())
 	}
 }
