+++ packages/server/src/checkin/output.rs
@@ -1,366 +1,285 @@
-use super::{
-	input,
-	object::{self, Metadata},
-};
-use crate::{Server, blob::create::Blob, temp::Temp};
-use futures::{FutureExt as _, TryStreamExt as _, stream::FuturesUnordered};
-use itertools::Itertools as _;
+use super::{Blob, File, State, Variant, set_permissions_and_times};
+use crate::{Server, temp::Temp};
+use bytes::Bytes;
+use futures::{TryStreamExt, stream::FuturesUnordered};
 use num::ToPrimitive;
-use std::{
-	collections::BTreeSet,
-	ffi::OsStr,
-	os::unix::fs::PermissionsExt as _,
-	path::PathBuf,
-	sync::{Arc, RwLock},
-};
+use std::{ops::Not, sync::Arc};
 use tangram_client as tg;
-use tangram_messenger::prelude::*;
-use tokio::task::JoinSet;
-
-#[derive(Debug)]
-pub struct Graph {
-	pub nodes: Vec<Node>,
-}
-
-#[derive(Clone, Debug)]
-pub struct Node {
-	pub blob: Option<Arc<Blob>>,
-	pub data: tg::artifact::Data,
-	pub edges: Vec<Edge>,
-	pub id: tg::artifact::Id,
-	pub input: Option<usize>,
-	pub metadata: Metadata,
-}
-
-#[derive(Clone, Debug)]
-pub struct Edge {
-	pub reference: tg::Reference,
-	pub subpath: Option<PathBuf>,
-	pub node: usize,
-}
-
-struct State {
-	nodes: Vec<Node>,
-	#[allow(clippy::option_option)]
-	visited: Vec<Option<Option<usize>>>,
-}
+use tangram_messenger::Messenger as _;
 
 impl Server {
-	pub(super) async fn create_output_graph(
+	pub(super) async fn checkin_cache_task(
 		&self,
-		input: &input::Graph,
-		object: &object::Graph,
-	) -> tg::Result<Arc<Graph>> {
-		// Create the state.
-		let state = RwLock::new(State {
-			nodes: Vec::with_capacity(object.nodes.len()),
-			visited: vec![None; object.nodes.len()],
-		});
-
-		// Recurse over the graph.
-		self.try_create_output_graph_inner(input, object, 0, &state)
-			.await?;
-
-		// Create the output.
-		let output = Graph {
-			nodes: state.into_inner().unwrap().nodes,
-		};
-
-		Ok(Arc::new(output))
-	}
-
-	async fn try_create_output_graph_inner(
-		&self,
-		input: &input::Graph,
-		object: &object::Graph,
-		object_index: usize,
-		state: &RwLock<State>,
-	) -> tg::Result<Option<usize>> {
-		// Check if this output node has been visited.
-		if let Some(output_index) = state.read().unwrap().visited[object_index] {
-			return Ok(output_index);
-		}
-
-		// Get the corresponding input node.
-		let input_index = object.nodes[object_index]
-			.unify
-			.object
-			.as_ref()
-			.left()
-			.copied();
-
-		// Create the output node.
-		let output_index = {
-			let mut state = state.write().unwrap();
-			let output_index = state.nodes.len();
-			let Ok(id) = object.nodes[object_index].id.clone().unwrap().try_into() else {
-				state.visited[object_index].replace(None);
-				return Ok(None);
-			};
-
-			let node = Node {
-				blob: object.nodes[object_index].blob.clone(),
-				input: input_index,
-				id,
-				data: object.nodes[object_index].data.clone().unwrap(),
-				metadata: object.nodes[object_index].metadata.clone().unwrap(),
-				edges: Vec::new(),
-			};
-			state.nodes.push(node);
-			state.visited[object_index].replace(Some(output_index));
-			output_index
-		};
-
-		// Create the output edges.
-		let edges: Vec<_> = object.nodes[object_index]
-			.edges
+		state: Arc<State>,
+		touched_at: i64,
+	) -> tg::Result<()> {
+		state
+			.graph
+			.roots
 			.iter()
-			.map(|edge| {
-				Box::pin(self.try_create_output_graph_inner(input, object, edge.index, state)).map(
-					|node| {
-						node.map(|index| {
-							index.map(|index| Edge {
-								reference: edge.reference.clone(),
-								subpath: edge.subpath.clone(),
-								node: index,
-							})
-						})
-					},
-				)
-			})
-			.collect::<FuturesUnordered<_>>()
-			.try_collect()
-			.await?;
-		let edges = edges.into_iter().flatten().collect();
+			.map(|(root, nodes)| {
+				let server = self.clone();
+				let root = *root;
+				let nodes = nodes.clone();
+				let state = state.clone();
+				async move {
+					// Copy the root artifact to the cache directory.
+					let semaphore = self.file_descriptor_semaphore.acquire().await.unwrap();
+					tokio::task::spawn_blocking({
+						let server = server.clone();
+						let state = state.clone();
+						move || server.checkin_copy_or_move_to_cache_directory(&state, root, &nodes)
+					})
+					.await
+					.unwrap()
+					.map_err(|source| tg::error!(!source, "failed to copy root object to temp"))?;
+					drop(semaphore);
 
-		// Update the output edges.
-		state.write().unwrap().nodes[output_index].edges = edges;
-
-		// Return the created node.
-		Ok(Some(output_index))
-	}
-
-	pub async fn checkin_cache_task_old(
-		&self,
-		output: &Graph,
-		input: &input::Graph,
-	) -> tg::Result<()> {
-		let mut stack = vec![0];
-		let mut visited = vec![false; output.nodes.len()];
-		let mut join_set = JoinSet::new();
-		while let Some(output_index) = stack.pop() {
-			if visited[output_index] {
-				continue;
-			}
-			visited[output_index] = true;
-			let output = &output.nodes[output_index];
-			stack.extend(output.edges.iter().map(|edge| edge.node));
-			let Some(input_index) = output.input else {
-				continue;
-			};
-			let Some(_) = &output.blob else {
-				continue;
-			};
-			let input = &input.nodes[input_index];
-			let executable = input.metadata.permissions().mode() & 0o111 != 0;
-			let path = input.arg.path.clone();
-			let cache_path = self.cache_path().join(output.id.to_string());
-			let temp = Temp::new(self);
-			join_set.spawn(async move {
-				if !tokio::fs::try_exists(&cache_path).await.map_err(|source| {
-					tg::error!(!source, "failed to check if the blob path exists")
-				})? {
-					tokio::fs::copy(path, temp.path())
-						.await
-						.map_err(|source| tg::error!(!source, "failed to copy the file"))?;
-					let mode = if executable { 0o755 } else { 0o644 };
-					let permissions = std::fs::Permissions::from_mode(mode);
-					tokio::fs::set_permissions(temp.path(), permissions)
-						.await
-						.map_err(|source| tg::error!(!source, "failed to set the permissions"))?;
-					tokio::fs::rename(temp.path(), cache_path)
-						.await
-						.map_err(|source| {
-							tg::error!(!source, "failed to rename the file to the blobs directory")
-						})?;
-				}
-				Ok::<_, tg::Error>(())
-			});
-		}
-		join_set
-			.join_all()
-			.await
-			.into_iter()
-			.try_collect::<_, (), _>()?;
-		Ok(())
-	}
-
-	pub async fn write_output_to_store(
-		&self,
-		output: Arc<Graph>,
-		object: Arc<object::Graph>,
-		touched_at: i64,
-	) -> tg::Result<()> {
-		let mut batch = Vec::new();
-
-		for (id, (data, _)) in &object.graphs {
-			batch.push((id.clone().into(), Some(data.serialize()?), None));
-		}
-
-		// Get the output in reverse-topological order.
-		let mut stack = vec![0];
-		let mut visited = vec![false; output.nodes.len()];
-		while let Some(output_index) = stack.pop() {
-			// Check if we've visited this node yet.
-			if visited[output_index] {
-				continue;
-			}
-			visited[output_index] = true;
-
-			// Get the output data.
-			let output = &output.nodes[output_index];
-
-			// Add the blob if the node has one.
-			if let Some(blob) = &output.blob {
-				let mut stack = vec![blob.as_ref()];
-				while let Some(blob) = stack.pop() {
-					if let Some(data) = &blob.data {
-						let bytes = data.serialize()?;
-						batch.push((blob.id.clone().into(), Some(bytes), None));
-					} else {
-						let reference = crate::store::CacheReference {
-							artifact: output.id.clone(),
-							subpath: None,
-							position: blob.position,
-							length: blob.length,
-						};
-						batch.push((blob.id.clone().into(), None, Some(reference)));
-					}
-					stack.extend(&blob.children);
-				}
-			}
-
-			batch.push((
-				output.id.clone().into(),
-				Some(output.data.serialize()?),
-				None,
-			));
-
-			stack.extend(output.edges.iter().map(|edge| edge.node));
-		}
-
-		let arg = crate::store::PutBatchArg {
-			objects: batch,
-			touched_at,
-		};
-		self.store
-			.put_batch(arg)
-			.await
-			.map_err(|source| tg::error!(!source, "failed to store the objects"))?;
-
-		Ok(())
-	}
-
-	pub async fn write_output_to_messenger(
-		&self,
-		output: Arc<Graph>,
-		object: Arc<object::Graph>,
-		touched_at: i64,
-	) -> tg::Result<()> {
-		let mut batch = Vec::new();
-
-		for (id, (data, _)) in &object.graphs {
-			batch.push((id.clone().into(), Some(data.serialize()?), None));
-
-			// Send the index object message.
-			let bytes = data.serialize()?;
-			let size = bytes.len().to_u64().unwrap();
-			let message = crate::index::Message::PutObject(crate::index::PutObjectMessage {
-				children: data.children(),
-				id: id.clone().into(),
-				size,
-				touched_at,
-				cache_reference: None,
-			});
-			let message = serde_json::to_vec(&message)
-				.map_err(|source| tg::error!(!source, "failed to serialize the message"))?;
-			let _published = self
-				.messenger
-				.stream_publish("index".to_owned(), message.into())
-				.await
-				.map_err(|source| tg::error!(!source, "failed to publish the message"))?;
-		}
-
-		// Get the output in reverse-topological order.
-		let mut stack = vec![0];
-		let mut visited = vec![false; output.nodes.len()];
-		while let Some(output_index) = stack.pop() {
-			// Check if we've visited this node yet.
-			if visited[output_index] {
-				continue;
-			}
-			visited[output_index] = true;
-
-			// Get the output data.
-			let output = &output.nodes[output_index];
-
-			// Add the blob if the node has one.
-			if let Some(blob) = &output.blob {
-				let mut stack = vec![blob.as_ref()];
-				while let Some(blob) = stack.pop() {
-					// Send the index message.
-					let children = blob
-						.data
+					// Publish a notification that the root is ready to be indexed.
+					let id = state.graph.nodes[root]
+						.object
 						.as_ref()
-						.map(tg::blob::Data::children)
-						.unwrap_or_default();
-					let id = blob.id.clone().into();
-					let size = blob.size;
+						.unwrap()
+						.id
+						.clone()
+						.try_into()
+						.unwrap();
 					let message =
-						crate::index::Message::PutObject(crate::index::PutObjectMessage {
-							cache_reference: Some(output.id.clone()),
-							children,
+						crate::index::Message::PutCacheEntry(crate::index::PutCacheEntryMessage {
 							id,
-							size,
 							touched_at,
 						});
 					let message = serde_json::to_vec(&message)
 						.map_err(|source| tg::error!(!source, "failed to serialize the message"))?;
-					let _published = self
+					let _published = server
 						.messenger
 						.stream_publish("index".to_owned(), message.into())
 						.await
 						.map_err(|source| tg::error!(!source, "failed to publish the message"))?;
-					stack.extend(&blob.children);
+					Ok::<_, tg::Error>(())
+				}
+			})
+			.collect::<FuturesUnordered<_>>()
+			.try_collect::<()>()
+			.await?;
+		Ok(())
+	}
+
+	fn checkin_copy_or_move_to_cache_directory(
+		&self,
+		state: &State,
+		root: usize,
+		nodes: &[usize],
+	) -> tg::Result<()> {
+		// Extract the root data.
+		let root_id = &state.graph.nodes[root].object.as_ref().unwrap().id;
+		let root_path = state.graph.nodes[root].path();
+		let root_metadata = state.graph.nodes[root].metadata.as_ref().unwrap();
+		let root_dst = self.cache_path().join(root_id.to_string());
+
+		// Attempt to rename if this is a destructive checkin.
+		if state.arg.destructive {
+			match std::fs::rename(root_path, &root_dst) {
+				Ok(()) => {
+					let epoch =
+						filetime::FileTime::from_system_time(std::time::SystemTime::UNIX_EPOCH);
+					filetime::set_symlink_file_times(&root_dst, epoch, epoch).map_err(
+						|source| tg::error!(!source, %path = root_dst.display(), "failed to set the modified time"),
+					)?;
+					return Ok(());
+				},
+				Err(ref error) if error.raw_os_error() == Some(libc::EEXIST | libc::ENOTEMPTY) => {
+					return Ok(());
+				},
+				Err(ref error) if error.raw_os_error() == Some(libc::ENOSYS) => (),
+				Err(source) => return Err(tg::error!(!source, "failed to rename root")),
+			}
+		}
+
+		// Otherwise copy to a temp.
+		let temp = Temp::new(self);
+		if root_metadata.is_file() {
+			// Copy the file.
+			std::fs::copy(root_path, temp.path())
+				.map_err(|source| tg::error!(!source, "failed to copy file"))?;
+
+			// Update permissions.
+			set_permissions_and_times(temp.path(), root_metadata)?;
+		} else if root_metadata.is_symlink() {
+			// Get the target.
+			let target = &state.graph.nodes[root].variant.unwrap_symlink_ref().target;
+
+			// Create the symlink.
+			std::os::unix::fs::symlink(target, temp.path())
+				.map_err(|source| tg::error!(!source, "failed to create link"))?;
+		} else {
+			// Walk the subgraph.
+			std::fs::create_dir_all(temp.path())
+				.map_err(|source| tg::error!(!source, "failed to create temp directory"))?;
+
+			// Copy everything except the root.
+			for node in nodes {
+				let node = &state.graph.nodes[*node];
+				let metadata = node.metadata.as_ref().unwrap();
+				let src = node.path();
+				let dst = temp.path().join(src.strip_prefix(root_path).unwrap());
+				if metadata.is_dir() {
+					std::fs::create_dir_all(&dst).map_err(
+						|source| tg::error!(!source, %path = dst.display(), "failed to create directory"),
+					)?;
+				} else if metadata.is_file() {
+					std::fs::create_dir_all(dst.parent().unwrap()).map_err(
+						|source| tg::error!(!source, %path = dst.display(), "failed to create parent directory"),
+					)?;
+					std::fs::copy(src, &dst).map_err(
+						|source| tg::error!(!source, %path = src.display(), "failed to copy file"),
+					)?;
+				} else if metadata.is_symlink() {
+					let target = &node.variant.unwrap_symlink_ref().target;
+					std::os::unix::fs::symlink(target, &dst)
+						.map_err(|source| tg::error!(!source, "failed to create the symlink"))?;
+				} else {
+					unreachable!("well, fuck")
+				}
+
+				// Update permissions.
+				set_permissions_and_times(&dst, metadata)?;
+			}
+		}
+
+		// Rename the temp to the cache.
+		match std::fs::rename(temp.path(), &root_dst) {
+			Ok(()) => (),
+			Err(error)
+				if matches!(
+					error.kind(),
+					std::io::ErrorKind::AlreadyExists | std::io::ErrorKind::DirectoryNotEmpty,
+				) => {},
+			Err(error) => {
+				return Err(tg::error!(
+					!error,
+					"failed to rename the path to the cache directory"
+				));
+			},
+		}
+
+		// Set the file times to the epoch.
+		let epoch = filetime::FileTime::from_system_time(std::time::SystemTime::UNIX_EPOCH);
+		filetime::set_symlink_file_times(&root_dst, epoch, epoch).map_err(
+			|source| tg::error!(!source, %path = root_dst.display(), "failed to set the modified time"),
+		)?;
+
+		Ok(())
+	}
+
+	pub(super) async fn checkin_messenger_task(
+		&self,
+		state: &Arc<State>,
+		touched_at: i64,
+	) -> tg::Result<()> {
+		if self.messenger.is_left() {
+			self.checkin_messenger_memory(state, touched_at).await?;
+		} else {
+			self.checkin_messenger_nats(state, touched_at).await?;
+		}
+		Ok(())
+	}
+
+	async fn checkin_messenger_memory(
+		&self,
+		state: &Arc<State>,
+		touched_at: i64,
+	) -> tg::Result<()> {
+		let mut messages: Vec<Bytes> = Vec::new();
+		for (root, nodes) in &state.graph.roots {
+			let nodes = std::iter::once(*root).chain(nodes.iter().copied());
+			let root: tg::artifact::Id = state.graph.nodes[*root]
+				.object
+				.as_ref()
+				.unwrap()
+				.id
+				.clone()
+				.try_into()
+				.unwrap();
+			for node in nodes {
+				let node = &state.graph.nodes[node];
+				let object = node.object.as_ref().unwrap();
+				let message = crate::index::Message::PutObject(crate::index::PutObjectMessage {
+					children: object.data.children(),
+					id: object.id.clone(),
+					size: object.bytes.len().to_u64().unwrap(),
+					touched_at,
+					cache_reference: None,
+				});
+				let message = serde_json::to_vec(&message)
+					.map_err(|source| tg::error!(!source, "failed to serialize the message"))?;
+				messages.push(message.into());
+				// Send messages for the blob.
+				if let Variant::File(File {
+					blob: Some(Blob::Create(blob)),
+					..
+				}) = &node.variant
+				{
+					let mut stack = vec![blob];
+					while let Some(blob) = stack.pop() {
+						let children = blob
+							.data
+							.as_ref()
+							.map(tg::blob::Data::children)
+							.unwrap_or_default();
+						let id = blob.id.clone().into();
+						let size = blob.size;
+						let message =
+							crate::index::Message::PutObject(crate::index::PutObjectMessage {
+								cache_reference: Some(root.clone()),
+								children,
+								id,
+								size,
+								touched_at,
+							});
+						let message = serde_json::to_vec(&message).map_err(|source| {
+							tg::error!(!source, "failed to serialize the message")
+						})?;
+						messages.push(message.into());
+						stack.extend(&blob.children);
+					}
 				}
 			}
-
-			// Send the index object message.
-			let bytes = output.data.serialize()?;
-			let size = bytes.len().to_u64().unwrap();
-			let message = crate::index::Message::PutObject(crate::index::PutObjectMessage {
-				children: output.data.children(),
-				id: output.id.clone().into(),
-				size,
-				touched_at,
-				cache_reference: None,
-			});
-			let message = serde_json::to_vec(&message)
-				.map_err(|source| tg::error!(!source, "failed to serialize the message"))?;
-			let _published = self
+		}
+		while !messages.is_empty() {
+			let published = self
 				.messenger
-				.stream_publish("index".to_owned(), message.into())
+				.stream_batch_publish("index".to_owned(), messages.clone())
 				.await
-				.map_err(|source| tg::error!(!source, "failed to publish the message"))?;
+				.map_err(|source| tg::error!(!source, "failed to publish the messages"))?
+				.await
+				.map_err(|source| tg::error!(!source, "failed to publish the messages"))?;
+			messages = messages.split_off(published.len());
+		}
+		Ok(())
+	}
 
-			// Send a message for the cache entry.
-			if output.blob.is_some() {
-				let id = output.id.clone();
-				let message =
-					crate::index::Message::PutCacheEntry(crate::index::PutCacheEntryMessage {
-						id,
-						touched_at,
-					});
+	async fn checkin_messenger_nats(&self, state: &Arc<State>, touched_at: i64) -> tg::Result<()> {
+		for (root, nodes) in &state.graph.roots {
+			let nodes = std::iter::once(*root).chain(nodes.iter().copied());
+			let root: tg::artifact::Id = state.graph.nodes[*root]
+				.object
+				.as_ref()
+				.unwrap()
+				.id
+				.clone()
+				.try_into()
+				.unwrap();
+			for node in nodes {
+				let node = &state.graph.nodes[node];
+				let object = node.object.as_ref().unwrap();
+				let message = crate::index::Message::PutObject(crate::index::PutObjectMessage {
+					children: object.data.children(),
+					id: object.id.clone(),
+					size: object.bytes.len().to_u64().unwrap(),
+					touched_at,
+					cache_reference: None,
+				});
 				let message = serde_json::to_vec(&message)
 					.map_err(|source| tg::error!(!source, "failed to serialize the message"))?;
 				let _published = self
@@ -368,19 +287,141 @@ impl Server {
 					.stream_publish("index".to_owned(), message.into())
 					.await
 					.map_err(|source| tg::error!(!source, "failed to publish the message"))?;
+
+				// Send messages for the blob.
+				if let Variant::File(File {
+					blob: Some(Blob::Create(blob)),
+					..
+				}) = &node.variant
+				{
+					let mut stack = vec![blob];
+					while let Some(blob) = stack.pop() {
+						let children = blob
+							.data
+							.as_ref()
+							.map(tg::blob::Data::children)
+							.unwrap_or_default();
+						let id = blob.id.clone().into();
+						let size = blob.size;
+						let message =
+							crate::index::Message::PutObject(crate::index::PutObjectMessage {
+								cache_reference: Some(root.clone()),
+								children,
+								id,
+								size,
+								touched_at,
+							});
+						let message = serde_json::to_vec(&message).map_err(|source| {
+							tg::error!(!source, "failed to serialize the message")
+						})?;
+						let _published = self
+							.messenger
+							.stream_publish("index".to_owned(), message.into())
+							.await
+							.map_err(|source| {
+								tg::error!(!source, "failed to publish the message")
+							})?;
+						stack.extend(&blob.children);
+					}
+				}
 			}
+		}
+		Ok(())
+	}
 
-			batch.push((
-				output.id.clone().into(),
-				Some(output.data.serialize()?),
-				None,
-			));
+	pub(super) async fn checkin_store_task(
+		&self,
+		state: &State,
+		touched_at: i64,
+	) -> tg::Result<()> {
+		let mut objects = Vec::with_capacity(state.graph.nodes.len());
 
-			stack.extend(output.edges.iter().map(|edge| edge.node));
+		// Add the graph objects.
+		for graph in &state.graph_objects {
+			objects.push((graph.id.clone().into(), Some(graph.bytes.clone()), None));
+			// Add all the graph children, too.
+			for (node, data) in graph.data.nodes.iter().enumerate() {
+				let (id, bytes) = match data.kind() {
+					tg::artifact::Kind::Directory => {
+						let data = tg::directory::data::Directory::Graph {
+							graph: graph.id.clone(),
+							node,
+						};
+						let bytes = data.serialize()?;
+						let id = tg::directory::Id::new(&bytes);
+						(id.into(), bytes)
+					},
+					tg::artifact::Kind::File => {
+						let data = tg::file::data::File::Graph {
+							graph: graph.id.clone(),
+							node,
+						};
+						let bytes = data.serialize()?;
+						let id = tg::file::Id::new(&bytes);
+						(id.into(), bytes)
+					},
+					tg::artifact::Kind::Symlink => {
+						let data = tg::symlink::data::Symlink::Graph {
+							graph: graph.id.clone(),
+							node,
+						};
+						let bytes = data.serialize()?;
+						let id = tg::symlink::Id::new(&bytes);
+						(id.into(), bytes)
+					},
+				};
+				objects.push((id, Some(bytes), None));
+			}
 		}
 
+		// Add the nodes.
+		for (root, nodes) in &state.graph.roots {
+			let nodes = std::iter::once(*root).chain(nodes.iter().copied());
+			let root_path = state.graph.nodes[*root].path();
+			let root: tg::artifact::Id = state.graph.nodes[*root]
+				.object
+				.as_ref()
+				.unwrap()
+				.id
+				.clone()
+				.try_into()
+				.unwrap();
+			for node in nodes {
+				let node = &state.graph.nodes[node];
+
+				// Add the object to the list to store.
+				let object = node.object.as_ref().unwrap();
+				objects.push((object.id.clone(), Some(object.bytes.clone()), None));
+
+				// If the file has a blob, add it too.
+				if let Variant::File(File {
+					blob: Some(Blob::Create(blob)),
+					..
+				}) = &node.variant
+				{
+					let mut stack = vec![blob];
+					while let Some(blob) = stack.pop() {
+						let subpath = node.path().strip_prefix(root_path).unwrap().to_owned();
+						let subpath = subpath
+							.to_str()
+							.unwrap()
+							.is_empty()
+							.not()
+							.then_some(subpath);
+						let reference = crate::store::CacheReference {
+							artifact: root.clone(),
+							subpath,
+							position: blob.position,
+							length: blob.length,
+						};
+						objects.push((blob.id.clone().into(), blob.bytes.clone(), Some(reference)));
+						stack.extend(&blob.children);
+					}
+				}
+			}
+		}
 		let arg = crate::store::PutBatchArg {
-			objects: batch,
+			objects,
 			touched_at,
 		};
 		self.store
@@ -391,402 +432,36 @@ impl Server {
 		Ok(())
 	}
 
-	/// Write hardlinks in the cache directory.
-	pub async fn write_links(
-		&self,
-		input: &input::Graph,
-		output: &Graph,
-		node: usize,
-	) -> tg::Result<()> {
-		let path = self.cache_path().join(output.nodes[node].id.to_string());
-		let mut visited = vec![false; output.nodes.len()];
-		self.write_links_inner(input, output, node, &path, &mut visited)
-			.await?;
-		Ok(())
-	}
-
-	async fn write_links_inner(
-		&self,
-		input: &input::Graph,
-		output: &Graph,
-		node: usize,
-		path: &PathBuf,
-		visited: &mut [bool],
-	) -> tg::Result<()> {
-		// Check if we've visited this node.
-		if visited[node] {
-			return Ok(());
-		}
-		visited[node] = true;
-
-		// macOS disables hardlinking for some files within .app dirs, such as *.app/Contents/{PkgInfo,Resources/\*.lproj,_CodeSignature} and .DS_Store. Perform a copy instead in these cases. See <https://github.com/NixOS/nix/blob/95f2b2beabc1ab0447bb4683516598265d71b695/src/libstore/optimise-store.cc#L100>.
-		let hardlink_prohibited = if cfg!(target_os = "macos") {
-			static APP_DIR_RE: std::sync::LazyLock<regex::Regex> =
-				std::sync::LazyLock::new(|| regex::Regex::new(r"\.app/Contents/.+$").unwrap());
-			APP_DIR_RE.is_match(path.to_string_lossy().as_ref())
-		} else {
-			false
-		};
-
-		// If this is a file, we need to create a hardlink in the cache directory and create a symlink for its contents in the blobs directory that points to the corresponding entry in the cache directory.
-		if let tg::artifact::Data::File(_) = &&output.nodes[node].data {
-			// Create hard link to the file or copy as needed.
-			let dst = self.cache_path().join(output.nodes[node].id.to_string());
-			if hardlink_prohibited {
-				let _permit = self.file_descriptor_semaphore.acquire().await.unwrap();
-				let result = tokio::fs::copy(path, &dst).await.map(|_| ());
-				match result {
-					Ok(()) => (),
-					Err(error) if error.raw_os_error() == Some(libc::EEXIST) => (),
-					Err(source) => {
-						let src = path.display();
-						let dst = dst.display();
-						return Err(tg::error!(!source, %src, %dst, "failed to copy"));
-					},
-				}
-			} else {
-				let result = tokio::fs::hard_link(path, &dst).await;
-				match result {
-					Ok(()) => (),
-					Err(error) if error.raw_os_error() == Some(libc::EEXIST) => (),
-					Err(source) => {
-						let src = path.display();
-						let dst = dst.display();
-						return Err(tg::error!(!source, %src, %dst, "failed to hardlink"));
-					},
-				}
-			}
-
-			// Since every file is a child of a directory we do not need to recurse over file dependencies and can bail early.
-			return Ok(());
-		}
-
-		// Recurse.
-		for edge in &output.nodes[node].edges {
-			// Skip any children that are roots.
-			let Some(input_index) = output.nodes[edge.node].input else {
-				continue;
-			};
-			if input.nodes[input_index].parent.is_none() {
-				continue;
-			}
-
-			// Get the path of the dependency.
-			let path_ = edge
-				.reference
-				.item()
-				.try_unwrap_path_ref()
-				.ok()
-				.or_else(|| edge.reference.options()?.path.as_ref())
-				.cloned()
-				.ok_or_else(|| tg::error!("expected a path dependency"))?;
-			let path_ = path.strip_prefix("./").unwrap_or(path_.as_ref());
-			// We could use metadata from the input, or the data, this avoids having to acquire a lock.
-			let path = if matches!(&output.nodes[node].id, tg::artifact::Id::Directory(_)) {
-				path.join(path_)
-			} else {
-				path.parent().unwrap().join(path_)
-			};
-
-			Box::pin(self.write_links_inner(input, output, edge.node, &path, visited)).await?;
-		}
-
-		Ok(())
-	}
-}
-
-impl Server {
-	pub(super) async fn copy_or_move_to_cache_directory(
-		&self,
-		input: &input::Graph,
-		output: &Graph,
-		node: usize,
-		progress: &crate::progress::Handle<tg::checkin::Output>,
-	) -> tg::Result<()> {
-		let mut stack = vec![node];
-		let mut visited = vec![false; output.nodes.len()];
-		while let Some(output_index) = stack.pop() {
-			// Check if this node has been visited.
-			if visited[output_index] {
-				continue;
-			}
-			visited[output_index] = true;
-
-			// Get the input index.
-			let Some(input_index) = output.nodes[output_index].input else {
-				continue;
-			};
-			if input.nodes[input_index].parent.is_none() {
-				// Create a temp.
-				let temp = Temp::new(self);
-
-				// Copy or move to the temp.
-				let mut visited = BTreeSet::new();
-				self.copy_or_move_all(
-					input,
-					output,
-					output_index,
-					temp.path().to_owned(),
-					&mut visited,
-					progress,
-				)
-				.await?;
-
-				// Update the file permissions.
-				self.update_permissions(input, output, output_index, temp.path().to_owned())
-					.await?;
-
-				// Rename to the cache directory.
-				let dest = self
-					.cache_path()
-					.join(output.nodes[output_index].id.to_string());
-				match tokio::fs::rename(&temp.path(), &dest).await {
-					Ok(()) => (),
-					Err(error) if error.raw_os_error() == Some(libc::EEXIST) => (),
-					Err(error) if error.raw_os_error() == Some(libc::ENOTEMPTY) => (),
-					Err(source) => return Err(tg::error!(!source, "failed to rename")),
-				}
-
-				// Update hardlinks.
-				self.write_links(input, output, output_index).await?;
-			}
-			// Recurse.
-			stack.extend(output.nodes[output_index].edges.iter().map(|e| e.node));
-		}
-
-		Ok(())
-	}
-
-	async fn copy_or_move_all(
-		&self,
-		input: &input::Graph,
-		output: &Graph,
-		node: usize,
-		dest: PathBuf,
-		visited: &mut BTreeSet<PathBuf>,
-		progress: &crate::progress::Handle<tg::checkin::Output>,
-	) -> tg::Result<()> {
-		// Check if we've visited this node.
-		let Some(input_index) = output.nodes[node].input else {
-			return Ok(());
-		};
-		let input_node = input.nodes[input_index].clone();
-		if visited.contains(&input_node.arg.path) {
-			return Ok(());
-		}
-		visited.insert(input_node.arg.path.clone());
-
-		if input_node.metadata.is_dir() {
-			if input_node.arg.destructive {
-				match tokio::fs::rename(&input_node.arg.path, &dest).await {
-					Ok(()) => return Ok(()),
-					Err(error) if error.raw_os_error() == Some(libc::ENOTEMPTY) => return Ok(()),
-					Err(error) if error.raw_os_error() == Some(libc::ENODEV) => (),
-					Err(source) => {
-						return Err(
-							tg::error!(!source, %src = input_node.arg.path.display(), %dst = dest.display(), "failed to rename directory"),
-						);
-					},
-				}
-			}
-			tokio::fs::create_dir_all(&dest)
-				.await
-				.map_err(|source| tg::error!(!source, "failed to create directory"))?;
-			let dependencies = output.nodes[node].edges.clone();
-			for edge in dependencies {
-				let Some(input_index) = output.nodes[edge.node].input else {
-					continue;
-				};
-				let input_node_ = &input.nodes[input_index];
-				let diff = input_node_
-					.arg
-					.path
-					.strip_prefix(&input_node.arg.path)
-					.unwrap();
-				let dest = dest.join(diff);
-				Box::pin(self.copy_or_move_all(input, output, edge.node, dest, visited, progress))
-					.await?;
-			}
-		} else if input_node.metadata.is_symlink() {
-			let target = self.get_symlink_target(input, output, node).await?;
-			tokio::fs::symlink(&target, &dest)
-				.await
-				.map_err(|source| tg::error!(!source, "failed to create symlink"))?;
-		} else if input_node.metadata.is_file() {
-			if input_node.arg.destructive {
-				match tokio::fs::rename(&input_node.arg.path, &dest).await {
-					Ok(()) => return Ok(()),
-					Err(error) if error.raw_os_error() == Some(libc::EEXIST) => return Ok(()),
-					Err(error) if error.raw_os_error() == Some(libc::ENODEV) => (),
-					Err(source) => {
-						return Err(
-							tg::error!(!source, %path = input_node.arg.path.display(), "failed to rename file"),
-						);
-					},
-				}
-			}
-			tokio::fs::copy(&input_node.arg.path, &dest).await.map_err(
-				|source| tg::error!(!source, %path = input_node.arg.path.display(), %dest = dest.display(), "failed to copy file"),
-			)?;
-		} else {
-			return Err(tg::error!(%path = input_node.arg.path.display(), "invalid file type"));
-		}
-
-		Ok(())
-	}
-
-	async fn update_permissions(
-		&self,
-		input: &input::Graph,
-		output: &Graph,
-		node: usize,
-		dest: PathBuf,
-	) -> tg::Result<()> {
-		let mut visited = vec![false; output.nodes.len()];
-		self.update_permissions_inner(input, output, node, dest, &mut visited)
-			.await
-	}
-
-	async fn update_permissions_inner(
-		&self,
-		input: &input::Graph,
-		output: &Graph,
-		node: usize,
-		dest: PathBuf,
-		visited: &mut [bool],
-	) -> tg::Result<()> {
-		// Check if we've visited this node.
-		if visited[node] {
-			return Ok(());
-		}
-		visited[node] = true;
-
-		match &output.nodes[node].data {
-			tg::artifact::Data::File(file) => {
-				let executable = match &file {
-					tg::file::Data::Graph { graph, node } => {
-						tg::Graph::with_id(graph.clone()).object(self).await?.nodes[*node]
-							.try_unwrap_file_ref()
-							.map_err(|_| tg::error!("expected a file"))?
-							.executable
-					},
-					tg::file::Data::Normal { executable, .. } => *executable,
-				};
-
-				let permissions = if executable { 0o0755 } else { 0o0644 };
-				tokio::fs::set_permissions(&dest, std::fs::Permissions::from_mode(permissions))
-					.await
-					.map_err(
-						|source| tg::error!(!source, %path = dest.display(), "failed to set file permissions"),
-					)?;
-				let metadata = tokio::fs::symlink_metadata(&dest).await.map_err(
-					|source| tg::error!(!source, %dest = dest.display(), "failed to get metadata"),
-				)?;
-				if !metadata.is_file() {
-					return Err(tg::error!(%path = dest.display(), "expected a file"));
-				}
-			},
-			tg::artifact::Data::Directory(_) => {
-				let permissions = tokio::fs::metadata(&dest)
-					.await
-					.map_err(
-						|source| tg::error!(!source, %dest = dest.display(), "failed to get metadata"),
-					)?
-					.permissions()
-					.mode();
-				let mode = format!("{permissions:o}");
-				tokio::fs::set_permissions(&dest, std::fs::Permissions::from_mode(0o0755))
-					.await
-					.map_err(
-						|source| tg::error!(!source, %path = dest.display(), %mode, "failed to set directory permissions"),
-					)?;
-			},
-			tg::artifact::Data::Symlink(_) => (),
-		}
-
-		// Recurse over directory entries.
-		if !matches!(
-			output.nodes[node].data.kind(),
-			tg::artifact::Kind::Directory
-		) {
-			return Ok(());
-		}
-		let dependencies = output.nodes[node].edges.clone();
-		let dependencies = dependencies
-			.into_iter()
-			.filter_map(|edge| {
-				let path = edge
-					.reference
-					.item()
-					.try_unwrap_path_ref()
-					.ok()
-					.or_else(|| edge.reference.options()?.path.as_ref())?;
-				Some((path.clone(), edge.node))
-			})
-			.collect::<Vec<_>>();
-
-		for (subpath, next_node) in dependencies {
-			let subpath = subpath.strip_prefix("./").unwrap_or(subpath.as_ref());
-			let dest = dest.join(subpath);
-			Box::pin(self.update_permissions_inner(input, output, next_node, dest, visited))
-				.await?;
-		}
-
-		Ok(())
-	}
-
-	async fn get_symlink_target(
-		&self,
-		input: &input::Graph,
-		output: &Graph,
-		symlink: usize,
-	) -> tg::Result<PathBuf> {
-		// Get the artifact and subpath, if they exist..
-		let artifact_and_subpath = output.nodes[symlink]
-			.edges
-			.first()
-			.map(|edge| (output.nodes[edge.node].id.clone(), edge.subpath.clone()))
-			.or_else(|| {
-				let tg::artifact::Data::Symlink(tg::symlink::Data::Artifact { artifact, subpath }) =
-					output.nodes[symlink].data.clone()
-				else {
-					return None;
-				};
-				Some((artifact, subpath))
-			});
-		if let Some((artifact, subpath)) = artifact_and_subpath {
-			// Find how deep this node is in the output tree.
-			let mut input_index = output.nodes[symlink].input.unwrap();
-			let mut depth = 0;
-			loop {
-				let Some(parent) = input.nodes[input_index].parent else {
-					break;
-				};
-				depth += 1;
-				input_index = parent;
-			}
-
-			// Create the target.
-			let id = artifact.to_string();
-			let subpath = subpath.unwrap_or_default();
-
-			let components = (0..depth)
-				.map(|_| std::path::Component::ParentDir)
-				.chain(std::iter::once(std::path::Component::Normal(OsStr::new(
-					&id,
-				))))
-				.chain(subpath.components());
-
-			return Ok(components.collect());
-		}
-
-		// Otherwise write the target directly.
-		if let tg::artifact::Data::Symlink(tg::symlink::Data::Target { target }) =
-			&output.nodes[symlink].data
+	pub(super) async fn checkin_create_lockfile_task(&self, state: &State) -> tg::Result<()> {
+		// Skip creating a lockfile if this is a destructive checkin, the caller did not request a lockfile, or the root is not a directory.
+		if state.arg.destructive
+			|| !state.arg.lockfile
+			|| !state.graph.nodes[0].metadata.as_ref().unwrap().is_dir()
 		{
-			return Ok(target.clone());
+			return Ok(());
 		}
 
-		Err(tg::error!(?symlink = &output.nodes[symlink], "invalid symlink"))
+		// Create a lockfile.
+		let lockfile = Self::create_lockfile(state)?;
+
+		// Do nothing if the lockfile is empty.
+		if lockfile.nodes.is_empty() {
+			return Ok(());
+		}
+
+		// Serialize the lockfile.
+		let contents = serde_json::to_vec_pretty(&lockfile)
+			.map_err(|source| tg::error!(!source, "failed to serialize lockfile"))?;
+
+		// Write to disk.
+		tokio::fs::write(
+			state.graph.nodes[0]
+				.path()
+				.join(tg::package::LOCKFILE_FILE_NAME),
+			contents,
+		)
+		.await
+		.map_err(|source| tg::error!(!source, "failed to write lockfile"))?;
+		Ok(())
 	}
 }
